refactor: split out classes into files
New classes:
- LifecycleWatcher
- ElementHandle
