+++ lib/Page.js
@@ -27,7 +27,7 @@ const Tracing = require('./Tracing');
 const {helper, debugError, assert} = require('./helper');
 const {Coverage} = require('./Coverage');
 const {Worker} = require('./Worker');
-const {createJSHandle} = require('./ExecutionContext');
+const {createJSHandle} = require('./ElementHandle');
 const {Accessibility} = require('./Accessibility');
 const writeFileAsync = helper.promisify(fs.writeFile);
