+++ lib/Worker.js
@@ -15,7 +15,8 @@
  */
 const EventEmitter = require('events');
 const {helper, debugError} = require('./helper');
-const {ExecutionContext, JSHandle} = require('./ExecutionContext');
+const {ExecutionContext} = require('./ExecutionContext');
+const {JSHandle} = require('./ElementHandle');
 
 class Worker extends EventEmitter {
   /**
